sui-sdk: remove dep on rocksdb
