+++ crates/sui-archival/src/lib.rs
@@ -29,10 +29,11 @@ use sui_config::genesis::Genesis;
 use sui_config::node::ArchiveReaderConfig;
 use sui_storage::blob::{Blob, BlobEncoding};
 use sui_storage::object_store::util::{get, put};
-use sui_storage::object_store::{ObjectStoreConfig, ObjectStoreGetExt, ObjectStorePutExt};
+use sui_storage::object_store::{ObjectStoreGetExt, ObjectStorePutExt};
 use sui_storage::{compute_sha3_checksum, SHA3_BYTES};
 use sui_types::base_types::ExecutionData;
 use sui_types::messages_checkpoint::{FullCheckpointContents, VerifiedCheckpointContents};
+use sui_types::storage::object_storage_config::ObjectStoreConfig;
 use sui_types::storage::{ReadStore, SingleCheckpointSharedInMemoryStore, WriteStore};
 use tracing::{error, info};
