+++ crates/sui-archival/src/tests.rs
@@ -18,10 +18,10 @@ use std::sync::Arc;
 use std::time::Duration;
 use sui_config::node::ArchiveReaderConfig;
 use sui_storage::object_store::util::path_to_filesystem;
-use sui_storage::object_store::{ObjectStoreConfig, ObjectStoreType};
 use sui_storage::{FileCompression, StorageFormat};
 use sui_swarm_config::test_utils::{empty_contents, CommitteeFixture};
 use sui_types::messages_checkpoint::{VerifiedCheckpoint, VerifiedCheckpointContents};
+use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
 use sui_types::storage::{ReadStore, SharedInMemoryStore, SingleCheckpointSharedInMemoryStore};
 use tempfile::tempdir;
