+++ crates/sui-config/Cargo.toml
@@ -24,7 +24,6 @@ prometheus.workspace = true
 narwhal-config.workspace = true
 sui-keys.workspace = true
 sui-protocol-config.workspace = true
-sui-storage.workspace = true
 sui-types.workspace = true
 workspace-hack.workspace = true
