+++ crates/sui-core/src/db_checkpoint_handler.rs
@@ -23,7 +23,7 @@ use sui_storage::object_store::util::{
     copy_recursively, find_all_dirs_with_epoch_prefix, find_missing_epochs_dirs,
     path_to_filesystem, put, run_manifest_update_loop, write_snapshot_manifest,
 };
-use sui_storage::object_store::{ObjectStoreConfig, ObjectStoreType};
+use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
 use tracing::{debug, error, info};
 use typed_store::rocks::MetricConf;
 
@@ -373,7 +373,7 @@ mod tests {
     use sui_storage::object_store::util::{
         find_all_dirs_with_epoch_prefix, find_missing_epochs_dirs, path_to_filesystem,
     };
-    use sui_storage::object_store::{ObjectStoreConfig, ObjectStoreType};
+    use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
     use tempfile::TempDir;
 
     #[tokio::test]
