+++ crates/sui-snapshot/src/writer.rs
@@ -26,8 +26,8 @@ use sui_core::authority::CHAIN_IDENTIFIER;
 use sui_protocol_config::{ProtocolConfig, ProtocolVersion};
 use sui_storage::blob::{Blob, BlobEncoding, BLOB_ENCODING_BYTES};
 use sui_storage::object_store::util::{copy_file, delete_recursively, path_to_filesystem};
-use sui_storage::object_store::ObjectStoreConfig;
 use sui_types::base_types::{ObjectID, ObjectRef};
+use sui_types::storage::object_storage_config::ObjectStoreConfig;
 use sui_types::sui_system_state::get_sui_system_state;
 use sui_types::sui_system_state::SuiSystemStateTrait;
 use tokio::sync::mpsc;
