+++ crates/sui-storage/src/object_store/http/mod.rs
@@ -10,7 +10,9 @@ use std::sync::Arc;
 use crate::object_store::http::gcs::GoogleCloudStorage;
 use crate::object_store::http::local::LocalStorage;
 use crate::object_store::http::s3::AmazonS3;
-use crate::object_store::{ObjectStoreConfig, ObjectStoreGetExt, ObjectStoreType};
+use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
+
+use crate::object_store::ObjectStoreGetExt;
 use anyhow::{anyhow, Context, Result};
 use chrono::{DateTime, Utc};
 use futures::{StreamExt, TryStreamExt};
@@ -122,9 +124,9 @@ fn header_meta(location: &Path, headers: &HeaderMap) -> Result<ObjectMeta> {
 #[cfg(test)]
 mod tests {
     use crate::object_store::http::HttpDownloaderBuilder;
-    use crate::object_store::{ObjectStoreConfig, ObjectStoreType};
     use object_store::path::Path;
     use std::fs;
+    use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
     use tempfile::TempDir;
 
     #[tokio::test]
