+++ crates/sui-storage/src/object_store/util.rs
@@ -414,10 +414,10 @@ mod tests {
     use crate::object_store::util::{
         copy_recursively, delete_recursively, write_snapshot_manifest, MANIFEST_FILENAME,
     };
-    use crate::object_store::{ObjectStoreConfig, ObjectStoreType};
     use object_store::path::Path;
     use std::fs;
     use std::num::NonZeroUsize;
+    use sui_types::storage::object_storage_config::{ObjectStoreConfig, ObjectStoreType};
     use tempfile::TempDir;
 
     #[tokio::test]
