+++ crates/sui-types/Cargo.toml
@@ -36,7 +36,9 @@ enum_dispatch.workspace = true
 eyre.workspace = true
 indexmap.workspace = true
 derivative.workspace = true
-
+clap.workspace = true
+object_store.workspace = true
+reqwest.workspace = true
 move-binary-format.workspace = true
 move-bytecode-utils.workspace = true
 move-command-line-common.workspace = true
