+++ crates/sui-types/src/storage/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod error;
+pub mod object_storage_config;
 mod object_store_trait;
 mod read_store;
 mod shared_in_memory_store;
